+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **_Breaking Change_**: add `AdvancedEvent`, remove `CascadeEvent`.
 - Feat: Add `FnHelper/Entry/CBC.Fn`. #5
 - Feat: move extensions from UniUtils to this project.
+- Feat: all `AdvancedEvent` allow listener with 0 parameter.
 
 ## v2.0.0